requirements.txt
# ============================================================================
# CORE DEPENDENCIES (Streamlit, Data Handling)
# ============================================================================
streamlit==1.51.0       
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
joblib==1.4.0           

# Plotting Libraries
matplotlib==3.8.4
plotly==5.19.0
plotly-express==0.4.1

# ============================================================================
# ML & PREPROCESSING LIBRARIES
# ============================================================================
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0

# Scikit-learn Extensions 
scikit-optimize==0.10.2
imbalanced-learn==0.12.2

# ============================================================================
# AI, MLOPS & TESTING
# ============================================================================
groq==0.8.0             
mlflow==2.13.0
statsmodels==0.14.2
shap==0.45.1
pytest==8.3.2           
pytest-cov==5.0.0       

# RAG / Vector Store
langchain==0.2.0
langchain-community==0.2.0
faiss-cpu==1.8.0        
sentence-transformers==2.7.0

# Deployment & File Support
fastapi==0.111.0
uvicorn[standard]==0.30.1
openpyxl                
pyarrow                 
requests==2.32.3
python-dotenv==1.0.1